refactor: replay report
