+++ packages/arex/.env.electron
@@ -1,1 +1,1 @@
-VITE_COMPANY_NAME="helloarex"
+VITE_COMPANY_NAME="trip"
