+++ packages/arex/src/i18n/locales/cn/components.json
@@ -105,6 +105,8 @@
     "executor": "执行人",
     "replayStartTime": "开始时间",
     "report": "报告",
+    "moreReport": "更多报告",
+    "latest": "最新",
     "basicInfo": "基本信息",
     "replayPassRate": "回放通过率",
     "passRate": "Case 通过率",
