+++ packages/arex/src/i18n/locales/en/common.json
@@ -88,5 +88,6 @@
   "and": " and",
   "or": " or",
   "notChrome": "The current browser is not Chrome, and some functions may not work properly",
-  "chromeVersionTooLow" : "Chrome version is too low, and some functions may not work properly, please upgrade to the latest version"
+  "chromeVersionTooLow" : "Chrome version is too low, and some functions may not work properly, please upgrade to the latest version",
+  "share": "Share"
 }
