+++ packages/arex/src/panes/Replay/AppHeader/RecordedCase/RecordedCaseDetail.tsx
