+++ packages/arex/src/panes/Replay/Replay.tsx
@@ -1,7 +1,6 @@
 import {
   ArexPaneFC,
   clearLocalStorage,
-  CollapseTable,
   setLocalStorage,
   useTranslation,
 } from '@arextest/arex-core';
@@ -12,22 +11,20 @@ import React, { useEffect, useMemo, useRef, useState } from 'react';
 
 import { APP_ID_KEY } from '@/constant';
 import { ApplicationService, ReportService } from '@/services';
-import { PlanStatistics } from '@/services/ReportService';
 import { useMenusPanes } from '@/store';
 import { decodePaneKey } from '@/store/useMenusPanes';
 
+import AppHeader from './AppHeader';
 import AppOwnersConfig, { AppOwnerConfigRef } from './AppOwnersConfig';
-import AppTitle from './AppTitle';
-import PlanItem from './PlanItem';
-import PlanReport from './PlanReport';
+import ReplayReport, { ReplayReportRef } from './ReplayReport';
 
 const ReplayPage: ArexPaneFC = (props) => {
   const { activePane } = useMenusPanes();
   const { t } = useTranslation('components');
 
   const [replayWrapperRef] = useAutoAnimate();
+  const replayReportRef = useRef<ReplayReportRef>(null);
 
-  const [selectedPlan, setSelectedPlan] = useState<PlanStatistics>();
   const { id: appId } = useMemo(() => decodePaneKey(props.paneKey), [props.paneKey]);
 
   useEffect(() => {
@@ -35,9 +32,8 @@ const ReplayPage: ArexPaneFC = (props) => {
     return () => clearLocalStorage(APP_ID_KEY);
   }, [activePane?.id]);
 
-  const [refreshDep, setRefreshDep] = useState<number>();
   const handleRefreshDep = () => {
-    setRefreshDep(new Date().getTime()); // 触发 ReplayTable 组件请求更新
+    replayReportRef.current?.getReportList(true);
   };
 
   const [firstQueryRecordCount, setFirstQueryRecordCount] = useState(true);
@@ -89,7 +85,7 @@ const ReplayPage: ArexPaneFC = (props) => {
             />
           )}
 
-          <AppTitle
+          <AppHeader
             appId={appId}
             appName={appInfo?.appName}
             readOnly={!hasOwner}
@@ -99,27 +95,7 @@ const ReplayPage: ArexPaneFC = (props) => {
             onQueryRecordCount={queryRecordCount}
           />
 
-          <CollapseTable
-            active={!!selectedPlan}
-            table={
-              <PlanReport
-                appId={appId}
-                refreshDep={refreshDep}
-                recordCount={recordCount}
-                onSelectedPlanChange={setSelectedPlan}
-              />
-            }
-            panel={
-              <PlanItem
-                appId={appId}
-                selectedPlan={selectedPlan}
-                readOnly={!hasOwner}
-                filter={(record) => !!record.totalCaseCount}
-                onRefresh={handleRefreshDep}
-                onDelete={() => setSelectedPlan(undefined)}
-              />
-            }
-          />
+          <ReplayReport ref={replayReportRef} appId={appId} readOnly={!hasOwner} />
         </>
       )}
       <AppOwnersConfig ref={appOwnerConfigRef} appId={appId} onAddOwner={getAppInfo} />
